fix: prevent re-renders
